Convert FASTQ to BAM

This function converts a FASTQ file to a BAM file using a dummy SAM format. It reads the FASTQ file, extracts the sequence identifiers and sequences, and creates a dummy SAM file.

Usage

fastq_to_bam(fastq_file, sam_file = "dummy.sam", reference = "chr1")

Arguments

fastq_file

A character string specifying the path to the FASTQ file.

sam_file

A character string specifying the path to the output SAM file. Default is "dummy.sam".

reference

A character string specifying the reference sequence name. Default is "chr1".

Details

The function reads the FASTQ file and extracts the sequence identifiers and sequences. It then creates a dummy SAM file with the following fields: - QNAME: Query name (read identifier) - FLAG: Bitwise flag (0 for unmapped) - RNAME: Reference sequence name - POS: Position on the reference sequence - MAPQ: Mapping quality (255 for unmapped) - CIGAR: CIGAR string (length of the sequence) - RNEXT: Reference name of the next read (dummy value "*") - PNEXT: Position of the next read (dummy value 0) - TLEN: Template length (dummy value 0) - SEQ: Sequence - QUAL: Quality scores

Examples

# Convert a FASTQ file to a BAM file
output_dir <- tempdir()
n <- 2
read_length <- 8
patient_id <- "test_patient"
fastq_file <- file.path(output_dir, paste0(patient_id, ".fastq"))
fill_fastq(patient_id, output_dir, n, read_length)
#> Creating a new FASTQ file.
#> Empty FASTQ file created at: /tmp/RtmphZCDpS/test_patient.fastq
#> Populated /tmp/RtmphZCDpS/test_patient.fastq with 2 reads.
fastq_to_bam(fastq_file, sam_file = "output.sam", reference = "chr1")
#> Dummy SAM file written to: output.sam