Function to generate a random sample of reads in FASTQ format

This function generates a specified number of random reads in FASTQ format. Each read consists of a sequence identifier, a random DNA sequence, a separator line, and quality scores.

Usage

rreads(patient_id, n, read_length = 8)

Arguments

patient_id

A character string representing the patient ID.

n

An integer specifying the number of reads to generate.

read_length

An integer specifying the length of each read. Default is 8. It should be multiple of 4.

Value

A character vector containing the generated FASTQ lines.

Details

The function generates random DNA sequences using the nucleotides A, T, G, and C. The quality scores are represented as ASCII characters, with 'I' indicating high quality. The output is formatted as a FASTQ file, which consists of four lines per read:

Line 1: Sequence identifier (e.g., @patient_id_read1)

Line 2: Random DNA sequence

Line 3: Separator (e.g., +)

Line 4: Quality scores

@seealso create_empty_fastq for creating an empty FASTQ file.

Examples

# Generate 5 random reads for patient ID "patient_123"
rreads("patient_123", n = 5)
#>  [1] "@patient_123_read1" "GATTGGTG"           "+"                 
#>  [4] "IIIIIIII"           "@patient_123_read2" "AACATTGG"          
#>  [7] "+"                  "IIIIIIII"           "@patient_123_read3"
#> [10] "CCTGATGG"           "+"                  "IIIIIIII"          
#> [13] "@patient_123_read4" "CATATGTC"           "+"                 
#> [16] "IIIIIIII"           "@patient_123_read5" "ATACCTGA"          
#> [19] "+"                  "IIIIIIII"          
# Generate 10 random reads for patient ID "patient_456" with a read length of 10
rreads("patient_456", n = 10, read_length = 10)
#>  [1] "@patient_456_read1"  "TTGTAAGTAC"          "+"                  
#>  [4] "IIIIIIIIII"          "@patient_456_read2"  "ACCGCCCGTG"         
#>  [7] "+"                   "IIIIIIIIII"          "@patient_456_read3" 
#> [10] "TGTGTCTAGT"          "+"                   "IIIIIIIIII"         
#> [13] "@patient_456_read4"  "TGATCCACAG"          "+"                  
#> [16] "IIIIIIIIII"          "@patient_456_read5"  "CTATTTGCTC"         
#> [19] "+"                   "IIIIIIIIII"          "@patient_456_read6" 
#> [22] "TTAGGCAAAA"          "+"                   "IIIIIIIIII"         
#> [25] "@patient_456_read7"  "TAAATCAGGA"          "+"                  
#> [28] "IIIIIIIIII"          "@patient_456_read8"  "CGGCTCAGAG"         
#> [31] "+"                   "IIIIIIIIII"          "@patient_456_read9" 
#> [34] "ATAAAAGATG"          "+"                   "IIIIIIIIII"         
#> [37] "@patient_456_read10" "CGCTATTGCG"          "+"                  
#> [40] "IIIIIIIIII"